requirements.txt
datasets==3.5.0
fitz==0.0.1.dev2
matplotlib==3.10.3
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
protobuf==6.30.2
python-dotenv==1.1.0
python_pptx==1.0.2
pytorch_crf==0.7.2
Requests==2.32.3
scikit_learn==1.6.1
sentence_transformers==4.1.0
torch==2.3.1
tqdm==4.67.1
transformers==4.51.3
